requirements.txt
playwright>=1.40.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
streamlit>=1.30.0
python-dotenv>=1.0.0
pydantic>=2.0.0
python-dateutil>=2.8.0
pytest>=7.4.0
pytest-asyncio>=0.23.0
